fix : sgの導入、データ取得部分の移行
