+++ my_profile/src/app/page.tsx
@@ -5,13 +5,18 @@ import { SkillsAndQualification } from "../components/SkillsAndQualification";
 import { Career } from "../components/Career";
 import { Working } from "../components/Working";
 import { Footer } from "../components/Footer";
+import { getTopContent } from "../libs/microcms";
 
-export default function Home() {
+export const dynamic = "force-static"; // 静的生成明示
+
+export default async function Home() {
+  const topData = await getTopContent();
+
   return (
     <>
       <Navbar />
       <main className="mt-[50px]">
-        <TopContent />
+        <TopContent topData={topData} />
         <Portfolio />
         <SkillsAndQualification />
         <Career />
