+++ my_profile/src/libs/microcms.ts
@@ -16,3 +16,9 @@ export const client = createClient({
   apiKey: process.env.NEXT_PUBLIC_MICROCMS_API_KEY,
   serviceDomain: process.env.NEXT_PUBLIC_MICROCMS_SERVICE_DOMAIN,
 });
+
+// トップコンテンツのデータ取得
+export const getTopContent = async (): Promise<TopContentProps> => {
+  const data = await client.get({ endpoint: "top-content" });
+  return data;
+};
