+++ my_profile/src/types/topContent.ts
@@ -1,0 +1,10 @@
+type TopContentProps = {
+  image: {
+    url: string;
+    height: number;
+    width: number;
+  };
+  job_category: string;
+  name: string;
+  description: string;
+};
